Rewrap and reformat all Markdown files
Since (some of) these files are shown to the user, it seems reasonable
to ensure that they are formatted consistently.

This PR is the result of running

```
for f in **/*.md(.); do
  comrak --width 80 --gfm --to commonmark $f > $f.tmp
  mv $f.tmp $f
done
```

in Zsh where `comrak` is the lovely Rust-based Markdown parser and
formatter from https://github.com/kivikakk/comrak.