Fix the crate clean working directory check
The check was back to front, checking that the crate root wasn’t in the
current directory, rather than that the current directory wasn’t in the
crate root.

Given a crate root like
/home/user/.cargo/registry/src/github.com-1ecc6299db9ec823/crate-1.0.0,
this meant that the clean command would complain when run from
directories like /home/user, which is not implausible (and is how I
detected it); one had to traverse to a different branch in the file
system tree like /tmp or /home/user/subdirectory. Also it meant that the
clean command would *not* complain when run from a subdirectory of the
crate root, when it should.

I also added a better message for when the assertion fails.