Load repo urls for all local identities
I've noticed when a "root trust id" signs a locally generated id
for the first time, the url field of the "trust proof" was always empty.

This problem is often hidden since urls are observed from proofs.
When setting up a trusted root chain for an organization,
the trust proofs repo can still be empty though.

Fix it by reading the urls from all local ids instead of just the "current" one.