Do not fetch anything after `cargo crev trust`
I have just spent a bunch of time investigating a spooky error of
not being able to fetch unrelated URL, after I was signing a trust
proof.

Fetching after creating new trust proof is a bit too much automation.
It also delays the completion of the actual operation that was invoked.