fix: hangs on circular deps and `--dev-dependencies` flag
I investigated a case of hang during `cargo crev verify` and it
came down to a circular dependency caused by dev depencencies
between local crates in a workspace.

I decided to fix it by ignoring dev dependencies between local
crates while building the graph of dependencies from the Cargo's
`resolve`, which made me realize that I can make `--dev-dependencies`
/`--no-dev-dependencies` work, by filtering them conditionally
in the same place.

Every time I touch this areas it comes to bite me later, but the hang
is gone, and I tested `--dev-dependencies` in a couple of places
and it seems to work OK now.